pyproject.toml
[project]
name = "wildberries"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-qrcode>=0.3",
    "django-redis>=6.0.0",
    "django-rest-swagger>=2.2.0",
    "django-widget-tweaks>=1.5.0",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.9.1",
    "load-dotenv>=0.1.0",
    "pillow>=11.3.0",
    "psycopg>=3.2.10",
    "psycopg2-binary>=2.9.10",
    "python-barcode>=0.16.1",
    "pyzbar>=0.1.9",
    "reportlab>=4.4.3",
    "treepoem>=3.27.1",
    "wget>=3.2",
]
